requirements/common.txt
alembic==1.0.5
bleach==3.1.0
blinker==1.4
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-Moment==0.6.0
Flask-PageDown==0.2.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
Pillow==5.4.1
python-dateutil==2.7.5
python-editor==1.0.3
selenium==3.141.0
six==1.12.0
SQLAlchemy==1.2.15
text-unidecode==1.2
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
